feat(rust-crane): use treefmt by default
Treefmt can handle all formatting and is configurable using
`treefmt.toml`. This is now used for alejandra and rustfmt. The
`treefmt` executable has been added to the devShell so that formatting
may be done with a single command, and the same command is used in the
pre-commit-hooks.

When changes are needed the error messages are less verbose (previously
*each* check would display on its own, but now there is only a single
`treefmt` check line), but treefmt should format the files properly and
they will just need to be added using git. This should be acceptable
because it isn't strictly relevant which formatter failed since all of
the issues can be resolved by just running `treefmt` and adding the
changes.