fix(rust-crane): Make template work with latest packages
The template didn't work correctly with the latest version of
pre-commit-hooks.nix due to it packaging `treefmt` and causing a
conflict. The fix is giving the pre-commit-hooks flake-parts
module a package to use for `treefmt` instead of the default one.

This also changes cargo workspace dependencies to use the extended
syntax `dep = { workspace = true }` instead of `dep.workspace = true`.
It seems that much of the tooling does not like the shorthand (and there
were typos in the template).